scr_web/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2
python-multipart==0.0.6
pymongo==4.6.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles
google-generativeai==0.3.2
pydantic==2.5.2
numpy>=1.24.0
Pillow>=10.0.0
httpx==0.24.1
authlib==1.2.1
itsdangerous==2.1.2
email-validator==2.0.0